+++ package.json
@@ -2,21 +2,21 @@
   "name": "gym-api",
   "version": "1.0.0",
   "description": "",
-  "main": "src/server.js",
+  "type": "module",
+  "main": "src/index.js",
   "scripts": {
-    "dev": "nodemon src/server.js",
+    "dev": "nodemon src/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "module",
   "dependencies": {
+    "bcrypt": "^6.0.0",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
-    "mongoose": "^8.16.0",
-    "nodemon": "^3.1.10",
-    "bcrypt": "^6.0.0"
-    
+    "jsonwebtoken": "^9.0.2",
+    "mongoose": "^8.15.2",
+    "nodemon": "^3.1.10"
   }
 }
