+++ src/app.js
@@ -1,0 +1,18 @@
+import express from 'express';
+import dotenv from 'dotenv';
+import userRoutes from './routes/userRoutes.js';
+import exerciseRoutes from './routes/exerciseRoutes.js';
+import workoutRoutes from './routes/workoutRoutes.js';
+import authRoutes from './routes/authRoutes.js';
+
+dotenv.config();
+
+const app = express();
+app.use(express.json());
+
+app.use('/api/auth', authRoutes);
+app.use('/api/users', userRoutes);
+app.use('/api/exercises', exerciseRoutes);
+app.use('/api/workouts', workoutRoutes);
+
+export default app;
