+++ src/controllers/authController.js
@@ -1,4 +1,5 @@
 import User from '../models/User.js';
+import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 
 export const register = async (req, res) => {
@@ -11,5 +12,20 @@ export const register = async (req, res) => {
     res.status(201).json({ token });
   } catch (err) {
     res.status(400).json({ error: err.message });
+  }
+};
+
+export const login = async (req, res) => {
+  try {
+    const { email, senha } = req.body;
+    const user = await User.findOne({ email });
+    if (!user || !(await bcrypt.compare(senha, user.senha))) {
+      return res.status(401).json({ error: 'Credenciais inválidas' });
+    }
+
+    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, { expiresIn: '7d' });
+    res.json({ token });
+  } catch (err) {
+    res.status(500).json({ error: err.message });
   }
 };
