+++ src/routes/authRoutes.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import { register, login } from '../controllers/authController.js';
+
+const router = express.Router();
+router.post('/register', register);
+router.post('/login', login);
+
+export default router;
