+++ src/server.js
@@ -1,22 +1,8 @@
-import express from 'express';
-import mongoose from 'mongoose';
-import dotenv from 'dotenv';
-import userRoutes from './routes/userRoutes.js';
-import exerciseRoutes from './routes/exerciseRoutes.js';
-import workoutRoutes from './routes/workoutRoutes.js';
+import app from "./app.js";
+import mongoose from "mongoose";
 
-dotenv.config();
-
-const app = express();
-app.use(express.json());
-
-mongoose.connect(process.env.MONGODB_URI)
-  .then(() => console.log('MongoDB conectado'))
-  .catch((err) => console.error('Erro ao conectar ao MongoDB:', err));
-
-app.use('/api/users', userRoutes);
-app.use('/api/exercises', exerciseRoutes);
-app.use('/api/workouts', workoutRoutes);
-
-const PORT = 3000;
-app.listen(PORT, () => console.log(`Servidor rodando na porta ${PORT}`));
+mongoose.connect(process.env.MONGODB_URI).then(() => {
+  console.log("MongoDB conectado");
+  const PORT = 3000;
+  app.listen(PORT, () => console.log(`Servidor rodando na porta ${PORT}`));
+});
